+++ src/raw/io/output/entry.rs
