+++ src/raw/io/output/mod.rs
@@ -42,11 +42,11 @@ mod into_streamable;
 pub use into_streamable::into_streamable_reader;
 pub use into_streamable::IntoStreamableReader;
 
-mod object_entry;
-pub use object_entry::ObjectEntry;
+mod entry;
+pub use entry::ObjectEntry;
 
-mod object_page;
-pub use object_page::BlockingObjectPage;
-pub use object_page::BlockingObjectPager;
-pub use object_page::ObjectPage;
-pub use object_page::ObjectPager;
+mod page;
+pub use page::BlockingObjectPage;
+pub use page::BlockingObjectPager;
+pub use page::ObjectPage;
+pub use page::ObjectPager;
